+++ CPU16/bases.h
@@ -1,9 +1,9 @@
-void hex32(int);
-void hex16(int);
-void hex8(int);
-void hex4(int);
+void hex32(uint16_t);
+void hex16(uint16_t);
+void hex8(uint16_t);
+void hex4(uint16_t);
 
-void bin32(int);
-void bin8(int);
-void bin16(int);
-void bin4(int);
+void bin32(uint16_t);
+void bin8(uint16_t);
+void bin16(uint16_t);
+void bin4(uint16_t);
