+++ CPU16/bit_tools.c
@@ -1,6 +1,6 @@
 
-unsigned int set_bit_on(int reg, int pos){
-    int mask, output;
+uint16_t set_bit_on(uint16_t reg, uint16_t pos){
+    uint16_t mask, output;
     
     mask = 1 << pos;
     output = reg | mask;
@@ -9,8 +9,8 @@ unsigned int set_bit_on(int reg, int pos){
 }
 
 
-unsigned int set_bit_off(int reg, int pos){
-    unsigned int mask, output;
+uint16_t set_bit_off(uint16_t reg, uint16_t pos){
+    uint16_t mask, output;
     
     mask = ~(1 << pos);
     output = reg & mask;
@@ -18,8 +18,8 @@ unsigned int set_bit_off(int reg, int pos){
     return output;
 }
 
-unsigned int get_bit(int reg, int pos){
-    unsigned int mask, output;
+uint16_t get_bit(uint16_t reg, uint16_t pos){
+    uint16_t mask, output;
     
     mask = 1;
     output = (reg >> pos) & mask;
@@ -27,8 +27,8 @@ unsigned int get_bit(int reg, int pos){
     return output;
 }
 
-unsigned int flip_one_bit(int reg, int pos){
-    int mask, bit, output;
+uint16_t flip_one_bit(uint16_t reg, uint16_t pos){
+    uint16_t mask, bit, output;
     
     bit = get_bit(reg,pos);
     if (bit == 0){
@@ -42,8 +42,8 @@ unsigned int flip_one_bit(int reg, int pos){
     return output;
 }
 
-int twos_complement(int A){
-    int output;
+int16_t twos_complement(uint16_t A){
+    uint16_t output;
     
     output = ~(A) + 1;
     
@@ -56,8 +56,8 @@ int twos_complement(int A){
  * a 32-bit int where all the bits are turned on
  */
  
-unsigned int fill_register (int val){
-    int output, val_bit;
+uint16_t fill_register (uint16_t val){
+    uint16_t output, val_bit;
     
     val_bit = get_bit(val,0);       //just to make sure that the argument passed is treated as a single bit
     output = twos_complement(val_bit);
@@ -67,15 +67,15 @@ unsigned int fill_register (int val){
 
 
 //returns most significant bit
-unsigned int get_msb(int reg){
-    unsigned int output;
+uint16_t get_msb(uint16_t reg){
+    uint16_t output;
     
     output = get_bit(reg,31);
     return output;
 }
 
-unsigned int get_lsb(int reg){
-    unsigned int output;
+uint16_t get_lsb(uint16_t reg){
+    uint16_t output;
     
     output = get_bit(reg,0);
     return output;
